+++ src/app/lib/widgets/fhir-search-dlg/fhir-search-dlg.component.ts
@@ -1,11 +1,15 @@
 import {Component, inject} from '@angular/core';
 import {FhirService, FHIRServer} from '../../../services/fhir.service';
-import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
+import {NgbActiveModal, NgbTooltipModule} from '@ng-bootstrap/ng-bootstrap';
 import fhir from 'fhir/r4';
 import {fhirPrimitives} from '../../../fhir';
-import {BehaviorSubject, Observable, Subject} from 'rxjs';
-import {map, switchMap, tap} from 'rxjs/operators';
-import { faInfoCircle } from '@fortawesome/free-solid-svg-icons';
+import {BehaviorSubject, Observable, ObservableInput, of, Subject} from 'rxjs';
+import {catchError, finalize, map, switchMap, tap} from 'rxjs/operators';
+import { faInfoCircle, faExclamationTriangle } from '@fortawesome/free-solid-svg-icons';
+import {FHIR_VERSION_TYPE} from "../../util";
+import {FormsModule} from "@angular/forms";
+import {CommonModule} from "@angular/common";
+import {FontAwesomeModule} from "@fortawesome/angular-fontawesome";
 // Search related inputs on the page.
 interface State {
   searchTerm: string;
@@ -17,10 +21,11 @@ interface SearchField {
   field: string,
   display: string,
   searchFieldPlaceholder: string
+  supportedVersions?: FHIR_VERSION_TYPE []; // If not defined, assume all versions are supported.
 }
 @Component({
-  standalone: false,
   selector: 'lfb-fhir-search-dlg',
+  imports: [FormsModule, CommonModule, FontAwesomeModule, NgbTooltipModule],
   styles: [`
     .result-item:hover {
       background-color: lightgrey;
@@ -52,13 +57,17 @@ export class FhirSearchDlgComponent {
   prevUrl: fhirPrimitives.url = null;
   total: number = undefined;
   questionnaires: fhir.Questionnaire [];
+  searchError: {message?: string, details?: string[]};
+  errorDetailsShow = false;
 
   /**
    * Define a structure to associate fhir search field, its modifiers, display and search input field placeholder.
    */
-  searchFieldList: SearchField [] = [
+  _searchFieldList: SearchField [] = [
     {field: '_content', display: 'Any text field', searchFieldPlaceholder: 'Search any text field'},
-    {field: 'code', display: 'Item code', searchFieldPlaceholder: 'Search item code'},
+    {field: 'questionnaire-code', display: 'Questionnaire code', searchFieldPlaceholder: 'Search questionnaire code', supportedVersions: ['R5']},
+    {field: 'code', display: 'Item code', searchFieldPlaceholder: 'Search item code', supportedVersions: ['STU3', 'R4']},
+    {field: 'item-code', display: 'Item code', searchFieldPlaceholder: 'Search item code', supportedVersions: ['R5']},
     {field: 'title:contains', display: 'Form title only', searchFieldPlaceholder: 'Search form title'},
     {field: 'name:contains', display: 'Form name only', searchFieldPlaceholder: 'Search form name'}
   ];
@@ -70,16 +79,37 @@ export class FhirSearchDlgComponent {
    */
   private _state: State = {
     searchTerm: '',
-    searchField: this.searchFieldList[0],
+    searchField: this._searchFieldList[0],
     fhirServer: this.fhirService.getFhirServer()
   };
 
   constructor() {
     // Set up search pipeline
     this._search$.pipe(
-      tap(() => this._loading$.next(true)),
-      switchMap(() => this._search()),
-      tap(() => this._loading$.next(false))
+      tap(() => {
+        this.searchError = null;
+        this.errorDetailsShow = false;
+        this._loading$.next(true);
+      }),
+      switchMap(() => this._search().pipe(
+        catchError((error, caught): ObservableInput<any> => {
+          const errorList = JSON.parse(error.message.replace(/^.*\n(?={)/s, ''))?.issue
+
+          this.searchError = {details: errorList?.map(e => e.diagnostics)};
+          if (error.statusCode < 500 && error.statusCode >= 400) {
+            this.searchError.message = 'The search parameters may not be supported by the selected FHIR server.';
+          }
+          else if (error.statusCode >= 500) {
+            this.searchError.message = 'Server error - the selected FHIR server encountered an internal error.';
+          }
+          else {
+            this.searchError.message = 'Error searching questionnaires.';
+          }
+          console.error(this.searchError.message, error);
+          return of(null);
+        }),
+        finalize(() => this._loading$.next(false))
+      )),
     ).subscribe((bundle) => {
       this.total = undefined; // Reset total before processing bundle
       this._bundle$.next(bundle);
@@ -88,16 +118,17 @@ export class FhirSearchDlgComponent {
     // Set up bundle pipeline. Bundle could be invoked either by search or navigation.
     this.bundle$.pipe(map((bundle) => {
       this.questionnaires = null;
+      this.nextUrl = null;
+      this.prevUrl = null;
+
       if(!bundle) {
-        return null; // Might happen when initializing _bundle$
+        return null; // Might happen when initializing _bundle$ or when error occurs.
       }
       if(bundle.total !== undefined) { // page bundles may not have total. The existing total is valid.
         this.total = bundle.total;
       }
 
       // Capture navigation urls.
-      this.nextUrl = null;
-      this.prevUrl = null;
       if(bundle.link && bundle.link && bundle.link.length > 0) {
         bundle.link.forEach((lnk) => {
           switch (lnk.relation) {
@@ -143,9 +174,39 @@ export class FhirSearchDlgComponent {
   get searchField() { return this._state.searchField; }
   set searchField(searchField: SearchField) { this._set({searchField});}
   get selectedFHIRServer() {return this._state.fhirServer;}
+
+  /**
+   * Set selected FHIR server. Make sure the search field is compatible with the selected FHIR server version.
+   * @param fhirServer - Selected FHIR server.
+   */
   set selectedFHIRServer(fhirServer: FHIRServer) {
     this.fhirService.setFhirServer(fhirServer);
-    this._set({fhirServer});
+    // Make sure the search field is compatible with the selected FHIR server version.
+    let searchField = this.searchField;
+    if (searchField.supportedVersions && !searchField.supportedVersions.includes(fhirServer.version)) {
+      // 'code' and 'item-code' are similar search fields in different FHIR versions. Try to find a suitalbe param.
+      if (searchField.field === 'code' || searchField.field === 'item-code') {
+        searchField = this._searchFieldList.find((sField) => {
+          return sField.supportedVersions?.includes(fhirServer.version) && (sField.field === 'code' || sField.field === 'item-code');
+        });
+      }
+      else if(searchField.field === 'questionnaire-code') {
+        // Set the default for incompatible search field.
+        searchField = this._searchFieldList[0];
+      }
+    }
+    this._set({fhirServer, searchField});
+  }
+
+  /**
+   * Get filtered search field list based on selected FHIR server version, used to populate search field dropdown.
+   */
+  get searchFieldList(): SearchField [] {
+    // Filter search fields based on selected FHIR server version.
+    return this._searchFieldList.filter((searchField) => {
+      // If supportedVersions is not defined, assume all versions are supported.
+      return !searchField.supportedVersions || searchField.supportedVersions.includes(this.selectedFHIRServer.version);
+    });
   }
 
   /**
@@ -233,4 +294,6 @@ export class FhirSearchDlgComponent {
   getServerList(): FHIRServer [] {
     return this.fhirService.fhirServerList;
   }
+
+  protected readonly faExclamationTriangle = faExclamationTriangle;
 }