+++ tests/contained-value-set.spec.ts
@@ -21,7 +21,7 @@ test.describe('Contained resources table in form level page', async () => {
 
     await page.getByRole('button', { name: 'Add new ValueSet' }).click();
     const dialog = page.locator('mat-dialog-container');
-    await dialog.waitFor({ state: 'visible', timeout: 5000 });
+    await dialog.waitFor({ state: 'visible', timeout: 10000 });
 
     await expect(dialog).toBeVisible();